refactor: 점수, 결과 출력 ResultView로 분리
